requirements.txt
pandas
requests
python-telegram-bot==20.7
python-dotenv
openpyxl
pyTelegramBotAPI
